requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
Django==5.1.5
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
yarl==1.18.3
